Databases/lab2/requirements.txt
django-admin-multiple-choice-list-filter>=0.1.1
django-admin-rangefilter>=0.5.1
django-advanced-filters>=1.1.1
django-debug-toolbar>=2.1
Django>=2.2.6
